Rename plugin package and move actions to actual actions
