+++ examples/routing/src/App.tsx
@@ -1,4 +1,4 @@
-import { Link, Redirect, Route, Switch } from "@prodo/routing-plugin";
+import { Link, Redirect, Route, Switch } from "@prodo/route";
 import * as React from "react";
 import { redirectHome, redirectUser } from "./actions";
 import { model } from "./model";
@@ -11,7 +11,7 @@ const HomePage = model.connect(({ dispatch }) => () => (
 ));
 const UserPage = model.connect(({ watch, route, dispatch }) => () => (
   <div>
-    Hello {watch(route.path)}.
+    Hello {watch(route.params.username)}.
     <button onClick={() => dispatch(redirectHome)()}>Log Out</button>
   </div>
 ));
@@ -22,7 +22,7 @@ const App = () => (
     <Link to={{ path: "/test", params: { user: "tom" } }}>Click Me</Link>
     <Switch>
       <Route exact path="/user" component={UserPage} />
-      <Route exact path="/test" />}
+      <Route exact path="/test" />
       <Route component={HomePage} />
     </Switch>
   </div>
