+++ packages/core/src/model.ts
@@ -1,3 +1,4 @@
+import { action } from "./action";
 import { connect } from "./connect";
 import { createStore } from "./store";
 import { Dispatch, Model, ProdoPlugin, Watch } from "./types";
@@ -9,14 +10,6 @@ export const createModel = <State>(): Model<
   { state: State; dispatch: Dispatch; watch: Watch }
 > => {
   const plugins: Array<ProdoPlugin<any, any, any, any>> = [];
-
-  const action = <A extends any[]>(
-    func: (c: any) => (...args: A) => void,
-    name: string = "(anonymous action)",
-  ) => {
-    (func as any).__name = name;
-    return func;
-  };
 
   const model: Model<any, any, any, any> = {
     createStore: config => createStore(config, plugins),
