+++ packages/route-plugin/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@prodo/routing-plugin",
+  "name": "@prodo/route",
   "version": "0.0.1",
   "main": "lib/index.js",
   "license": "MIT",
