+++ packages/route-plugin/src/types.ts
@@ -1,0 +1,25 @@
+import { History } from "history";
+
+export interface RouteParams {
+  path: string;
+  params?: { [key: string]: string };
+}
+
+export const historySymbol = Symbol("@@routing/history");
+export const universeSymbol = Symbol("@@routing/universe");
+
+export interface Universe {
+  route: RouteParams;
+}
+
+export interface Routing {
+  [historySymbol]: History;
+  [universeSymbol]: Universe;
+  route: RouteParams;
+}
+
+export interface Config {
+  route: {
+    history: History;
+  };
+}
