+++ tsconfig.json
@@ -15,7 +15,7 @@
     { "path": "packages/firestore-plugin" },
     { "path": "packages/local-plugin" },
     { "path": "packages/random-plugin" },
-    { "path": "packages/routing-plugin" },
+    { "path": "packages/route-plugin" },
     { "path": "packages/stream-plugin" }
   ]
 }
